backend-fastapi/requirements.txt
fastapi==0.115.*
uvicorn==0.30.*
httpx==0.27.*
pydantic==2.9.*
kafka-python==2.0.*
protobuf>=5.26.0
grpcio>=1.62.0 
grpcio-tools>=1.62.0